requirements.txt
selene == 2.0.0rc7
pytest == 8.0.0
allure-pytest == 2.13.2
python-dotenv == 1.0.0